+++ examples/main/modules.tf
@@ -28,7 +28,6 @@ module "logs" {
 }
 
 module "cdn_frontdoor" {
-
   source  = "claranet/cdn-frontdoor/azurerm"
   version = "x.x.x"
 
@@ -42,7 +41,7 @@ module "cdn_frontdoor" {
 
   logs_destinations_ids = [
     module.logs.log_analytics_workspace_id,
-    module.logs.logs_storage_account_id
+    module.logs.logs_storage_account_id,
   ]
 
   endpoints = [
@@ -52,7 +51,7 @@ module "cdn_frontdoor" {
     {
       name    = "azure"
       enabled = false
-    }
+    },
   ]
 
   origin_groups = [
@@ -76,7 +75,7 @@ module "cdn_frontdoor" {
         protocol            = "Https"
         request_type        = "GET"
       }
-    }
+    },
   ]
 
   origins = [
@@ -91,15 +90,13 @@ module "cdn_frontdoor" {
       origin_group_name              = "contoso2"
       certificate_name_check_enabled = false
       host_name                      = "azure.contoso.com"
-    }
+    },
   ]
 
-  custom_domains = [
-    {
-      name      = "www"
-      host_name = "www.contoso.com"
-    }
-  ]
+  custom_domains = [{
+    name      = "www"
+    host_name = "www.contoso.com"
+  }]
 
   routes = [
     {
@@ -122,167 +119,165 @@ module "cdn_frontdoor" {
       patterns_to_match   = ["/contoso"]
       supported_protocols = ["Http", "Https"]
       rule_sets_names     = ["my_rule_set", "my_rule_set2"]
-    }
+    },
   ]
 
   rule_sets = [
     {
       name                 = "my_rule_set"
       custom_resource_name = "custom_rule"
-      rules = [
-        {
-          name                 = "redirect"
-          custom_resource_name = "myrulename"
-          order                = 1
-          actions = {
-            url_rewrite_action = {
-              source_pattern = "/"
-              destination    = "/contoso"
-            }
-          }
-          conditions = {
-            is_device_condition = {
-              operator     = "Equal"
-              match_values = ["Desktop"]
-            }
-          }
+
+      rules = [{
+        name                 = "redirect"
+        custom_resource_name = "myrulename"
+        order                = 1
+        actions = {
+          response_header_action = [
+            {
+              header_action = "Overwrite"
+              header_name   = "Access-Control-Allow-Origin"
+              value         = "https://www.foo.bar.fr"
+            },
+            {
+              header_action = "Overwrite"
+              header_name   = "Access-Control-Allow-Credentials"
+              value         = "true"
+            },
+            {
+              header_action = "Overwrite"
+              header_name   = "Access-Control-Allow-Headers"
+              value         = "Authorization, Content-Type, ocp-apim-subscription-key"
+            },
+            {
+              header_action = "Overwrite"
+              header_name   = "Access-Control-Allow-Methods"
+              value         = "POST,PUT,GET,DELETE,OPTIONS"
+            },
+          ]
+          url_rewrite_action = [{
+            source_pattern = "/"
+            destination    = "/contoso"
+          }]
         }
-      ]
+        conditions = {
+          is_device_condition = [{
+            operator     = "Equal"
+            match_values = ["Desktop"]
+          }]
+        }
+      }]
     },
     {
       name                 = "my_rule_set2"
       custom_resource_name = "custom_rule2"
-    }
+    },
   ]
 
-  firewall_policies = [
-    {
-      name                              = "test"
-      enabled                           = true
-      mode                              = "Prevention"
-      redirect_url                      = "https://www.contoso.com"
-      custom_block_response_status_code = 403
-      custom_block_response_body        = "PGh0bWw+CjxoZWFkZXI+PHRpdGxlPkhlbGxvPC90aXRsZT48L2hlYWRlcj4KPGJvZHk+CkhlbGxvIHdvcmxkCjwvYm9keT4KPC9odG1sPg=="
+  firewall_policies = [{
+    name                              = "test"
+    enabled                           = true
+    mode                              = "Prevention"
+    redirect_url                      = "https://www.contoso.com"
+    custom_block_response_status_code = 403
+    custom_block_response_body        = "PGh0bWw+CjxoZWFkZXI+PHRpdGxlPkhlbGxvPC90aXRsZT48L2hlYWRlcj4KPGJvZHk+CkhlbGxvIHdvcmxkCjwvYm9keT4KPC9odG1sPg=="
 
-      custom_rules = [
-        {
-          name                           = "Rule1"
-          enabled                        = true
-          priority                       = 1
-          rate_limit_duration_in_minutes = 1
-          rate_limit_threshold           = 10
-          type                           = "MatchRule"
-          action                         = "Block"
+    custom_rules = [
+      {
+        name                           = "Rule1"
+        enabled                        = true
+        priority                       = 1
+        rate_limit_duration_in_minutes = 1
+        rate_limit_threshold           = 10
+        type                           = "MatchRule"
+        action                         = "Block"
+        match_conditions = [{
+          match_variable     = "RemoteAddr"
+          operator           = "IPMatch"
+          negation_condition = false
+          match_values       = ["10.0.1.0/24", "10.0.0.0/24"]
+        }]
+      },
+      {
+        name                           = "Rule2"
+        enabled                        = true
+        priority                       = 2
+        rate_limit_duration_in_minutes = 1
+        rate_limit_threshold           = 10
+        type                           = "MatchRule"
+        action                         = "Block"
+        match_conditions = [
+          {
+            match_variable     = "RemoteAddr"
+            operator           = "IPMatch"
+            negation_condition = false
+            match_values       = ["192.168.1.0/24"]
+          },
+          {
+            match_variable     = "RequestHeader"
+            selector           = "UserAgent"
+            operator           = "Contains"
+            negation_condition = false
+            match_values       = ["windows"]
+            transforms         = ["Lowercase", "Trim"]
+          },
+        ]
+      },
+    ]
 
-          match_conditions = [
-            {
-              match_variable     = "RemoteAddr"
-              operator           = "IPMatch"
-              negation_condition = false
-              match_values       = ["10.0.1.0/24", "10.0.0.0/24"]
-            }
-          ]
-        },
-        {
-          name                           = "Rule2"
-          enabled                        = true
-          priority                       = 2
-          rate_limit_duration_in_minutes = 1
-          rate_limit_threshold           = 10
-          type                           = "MatchRule"
-          action                         = "Block"
-
-          match_conditions = [
-            {
-              match_variable     = "RemoteAddr"
-              operator           = "IPMatch"
-              negation_condition = false
-              match_values       = ["192.168.1.0/24"]
-            },
-            {
-              match_variable     = "RequestHeader"
-              selector           = "UserAgent"
-              operator           = "Contains"
-              negation_condition = false
-              match_values       = ["windows"]
-              transforms         = ["Lowercase", "Trim"]
-            }
-          ]
-        }
-      ]
-
-      managed_rules = [
-        {
-          type    = "DefaultRuleSet"
-          version = "1.0"
-          action  = "Log"
-
-          exclusions = [
-            {
+    managed_rules = [
+      {
+        type    = "DefaultRuleSet"
+        version = "1.0"
+        action  = "Log"
+        exclusions = [{
+          match_variable = "QueryStringArgNames"
+          operator       = "Equals"
+          selector       = "not_suspicious"
+        }]
+        overrides = [
+          {
+            rule_group_name = "PHP"
+            rules = [{
+              rule_id = "933100"
+              enabled = false
+              action  = "Block"
+            }]
+          },
+          {
+            rule_group_name = "SQLI"
+            exclusions = [{
               match_variable = "QueryStringArgNames"
               operator       = "Equals"
-              selector       = "not_suspicious"
-            }
-          ]
-
-          overrides = [
-            {
-              rule_group_name = "PHP"
-
-              rules = [
-                {
-                  rule_id = "933100"
-                  enabled = false
-                  action  = "Block"
-                }
-              ]
-            },
-            {
-              rule_group_name = "SQLI"
-
+              selector       = "really_not_suspicious"
+            }]
+            rules = [{
+              rule_id = "942200"
+              action  = "Block"
               exclusions = [{
                 match_variable = "QueryStringArgNames"
                 operator       = "Equals"
-                selector       = "really_not_suspicious"
-                }
-              ]
+                selector       = "innocent"
+              }]
+            }]
+          },
+        ]
+      },
+      {
+        type    = "Microsoft_BotManagerRuleSet"
+        version = "1.0"
+        action  = "Log"
+      },
+    ]
+  }]
 
-              rules = [{
-                rule_id = "942200"
-                action  = "Block"
-
-                exclusions = [
-                  {
-                    match_variable = "QueryStringArgNames"
-                    operator       = "Equals"
-                    selector       = "innocent"
-                  }
-                ]
-                }
-              ]
-            }
-          ]
-        },
-        {
-          type    = "Microsoft_BotManagerRuleSet"
-          version = "1.0"
-          action  = "Log"
-        }
-      ]
-
-    }
-  ]
-
-  security_policies = [
-    {
-      name                 = "MySecurityPolicy"
-      custom_resource_name = "MyBetterNamedSecurityPolicy"
-      firewall_policy_name = "test"
-      patterns_to_match    = ["/*"]
-      custom_domain_names  = ["www"]
-      endpoint_names       = ["web", "azure"]
-    }
-  ]
+  security_policies = [{
+    name                 = "MySecurityPolicy"
+    custom_resource_name = "MyBetterNamedSecurityPolicy"
+    firewall_policy_name = "test"
+    patterns_to_match    = ["/*"]
+    custom_domain_names  = ["www"]
+    endpoint_names       = ["web", "azure"]
+  }]
 
   extra_tags = {
     foo = "bar"